refactor: move the main scenario in its proper file
