+++ index.js
@@ -1,19 +1,7 @@
-import { Drug, Pharmacy } from "./pharmacy";
 import fs from "fs";
+import {run} from "./scenario";
 
-const drugs = [
-  new Drug("Doliprane", 20, 30),
-  new Drug("Herbal Tea", 10, 5),
-  new Drug("Fervex", 5, 40),
-  new Drug("Magic Pill", 15, 40)
-];
-const trial = new Pharmacy(drugs);
-
-const log = [];
-
-for (let elapsedDays = 0; elapsedDays < 30; elapsedDays++) {
-  log.push(JSON.stringify(trial.updateBenefitValue()));
-}
+const log = run();
 
 /* eslint-disable no-console */
 fs.writeFile("output.txt", log.toString(), err => {
